requirements.txt
aiogram
vk-api
supabase
python-dotenv
cryptography
Pillow
